+++ src/ng-vcl/embedded-input-group/embedded-input-group.component.ts
@@ -1,6 +1,8 @@
-import { Component, Output, EventEmitter, Input, AfterViewInit, ViewChild, HostBinding } from '@angular/core';
+import { Component, Output, EventEmitter, Input, AfterViewInit, ViewChild, HostBinding, ContentChildren, forwardRef, QueryList, Self, SkipSelf, Directive } from '@angular/core';
 import { NgModel } from '@angular/forms';
 import { ObservableComponent } from '../core/index';
+import { InputDirective } from '../input/index';
+import { ButtonDirective } from '../button/index';
 
 @Component({
   selector: 'vcl-embedded-input-group',
@@ -19,17 +21,67 @@ export class EmbeddedInputGroupComponent extends ObservableComponent {
   prepIcon?: string;
 
   @Input()
-  prepButtonIcon?: string;
+  appIcon?: string;
 
-  @Input()
-  appendIcon?: string;
+  @ContentChildren(forwardRef(() => EmbeddedButtonDirective))
+  buttons: QueryList<EmbeddedButtonDirective>;
 
-  @Input()
-  appButtonIcon?: string;
+  get hasAppendedItem() {
+    return !!this.appIcon || this.buttons.length > 0;
+  }
+}
 
-  @Output()
-  prepButtonTap: EventEmitter<Event> = new EventEmitter();
+@Directive({
+  selector: 'input[vcl-input][vcl-embedded-input-group]',
+})
+export class EmbeddedInputDirective {
+  constructor(
+    @Self() private input: InputDirective,
+    @SkipSelf() private inputGroup: EmbeddedInputGroupComponent
+  ) { }
 
-  @Output()
-  appButtonTap: EventEmitter<Event> = new EventEmitter();
+  get isDisabled() {
+    return this.input.disabled || this.inputGroup.disabled;
+  }
+
+  @HostBinding('class.vclDisabled')
+  get classDisabled() {
+    return this.isDisabled;
+  }
+
+  @HostBinding('attr.disabled')
+  get attrDisabled() {
+    return this.isDisabled ? true : null;
+  }
+
+  @HostBinding('class.vclPrepItem')
+  get prepItem() {
+    return !!this.inputGroup.prepIcon;
+  }
+
+  @HostBinding('class.vclAppItem')
+  get hasAppendedItem() {
+    return this.inputGroup.hasAppendedItem;
+  }
+}
+
+@Directive({
+  selector: 'button[vcl-button][vcl-embedded-input-group][icon]',
+  host: {
+    '[class.vclTransparent]': 'true',
+    '[class.vclAppended]': 'true'
+  }
+})
+export class EmbeddedButtonDirective {
+
+  constructor(
+    @Self() private button: ButtonDirective,
+    @SkipSelf() private inputGroup: EmbeddedInputGroupComponent
+  ) { }
+
+  @HostBinding('class.vclDisabled')
+  @HostBinding('attr.disabled')
+  get isDisabled() {
+    return (this.button.disabled || this.inputGroup.disabled) ? true : null;
+  }
 }