+++ src/ng-vcl/embedded-input-group/index.ts
@@ -3,14 +3,11 @@ import { CommonModule } from '@angular/common';
 import { VCLButtonModule } from '../button/index';
 import { VCLInputModule } from '../input/index';
 import { VCLIconModule } from '../icon/index';
-import { EmbeddedInputGroupComponent } from './embedded-input-group.component';
-import { EmbeddedInputDirective } from './embedded-input.directive';
-import { FormsModule } from '@angular/forms';
+import { EmbeddedInputGroupComponent, EmbeddedInputDirective, EmbeddedButtonDirective } from './embedded-input-group.component';
 
 @NgModule({
-  imports: [CommonModule, VCLButtonModule, VCLInputModule, VCLIconModule, FormsModule],
-  exports: [EmbeddedInputGroupComponent, EmbeddedInputDirective],
-  declarations: [EmbeddedInputGroupComponent, EmbeddedInputDirective],
-  providers: [],
+  imports: [CommonModule, VCLButtonModule, VCLInputModule, VCLIconModule],
+  exports: [EmbeddedInputGroupComponent, EmbeddedInputDirective, EmbeddedButtonDirective],
+  declarations: [EmbeddedInputGroupComponent, EmbeddedInputDirective, EmbeddedButtonDirective],
 })
 export class VCLEmbeddedInputGroupModule { }