+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - requires Angular ^6 / rxjs ^6
 - removed integrated l10n support
 - removed plotly.js & json-editor packages
+- embedded-input-group:
+  - removed prepButtonIcon/apppButtonIcon properties
+  - added ng-content for appended button
 - icon:
   - removed svguse/src properties
 - link: